backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
requests
python-multipart
sentence-transformers
faiss-cpu
textract
python-dotenv
bcrypt
